Zombies can now hide in the dark. Also player is more venerable to zombie attacks even when still
Spooky!
And please fix the rotation before i revert it to the keys. Try using
the steering AI i set up in this commit and just make the player's
location equal the mouse location. The steering AI will do the rest. Or
just get another AI like the zombie's to do everything.