Ballistic server branch (#10)
That was a pain, I'm not merging again
* fixed deltas, still needs work and testing

* Commit of random stuff before i perform datapacket transition

* Server update

* Made packets

* Finally! All data sent through packets

First step towards complex data sending

* Multiplayer going well

* I can't get more than one person to join

* Server can handle more than one player, client cant render them all

* Damn it feels good to be a gangsta

* Hold on there

Nothing is getting deprecated until i see a working server

* Bunch of fixes

Client now works without server.
Moved connection stuff all to player class

* Broken commit

* JSON!

* Server now supports multiple players and client can show them all!

Just need to clean up the mess the client makes when viewing them all

* Cleanup of error handling

* The impossible has been done

Smooth player enter and exit on client and server. 
I can't believe it finally works. Also fixed a bunch of other things
like hearts and fps counter

* Added google Gson which should be faster in sending json data

* game connects to server successfully 

now its time to do the big stuff on the server

* client knows when to disconnect.

Just to make it do something when it knows.

* Whoops

* Recommit: client knows when to disconnect.

Just to make it do something when it knows.

* for some reason my changes didnt go live

* Actually multiplayer.Lag will be fixed later

Literally took me 4 fucking hours of coding and debugging to get this to work

* fixed the lag which took a bit of time to figure out. Stupid mistake

* Working on scrolling bg for main menu

* Asset manager coming in slowly

A pain, can't event switch to master branch for this

* Fully moved to the Asset loader

For some reason the menu is weird. Very weird